+++ engine/include/graphics/renderer.hpp
@@ -27,7 +27,7 @@ namespace cppengine {
         static ModelID getModel(std::string const &name);
         static ShaderID getShader(std::string const &name);
 
-        static ModelID loadModel(std::string const &name, Vector3 const *vertices, std::size_t vertex_count, std::size_t const *indices, std::size_t index_count);
+        static ModelID loadModel(std::string const &name, Vector3 const *vertices, std::uint32_t vertex_count, std::uint32_t const *indices, std::uint32_t index_count);
         static ShaderID loadShader(std::string const &name);
         static Texture loadTexture(std::string const &path);
     };
