+++ engine/include/maths/box_2d.hpp
@@ -1,6 +1,8 @@
 #ifndef BOX_2D_HPP
 #define BOX_2D_HPP
 
+#include <cstdint>
+
 #include "maths/vector3.hpp"
 #include "maths/math_helper.hpp"
 
@@ -10,9 +12,9 @@ namespace cppengine {
     struct Box2D {
         using vertex_type = Vector3;
         using value_type = typename std::remove_reference<decltype(*(& std::declval<vertex_type&>()[0]))>::type;
-        static constexpr std::size_t vertex_count = 4;
-        static constexpr std::size_t index_count = 6;
-        static constexpr std::size_t indices[index_count] = { 0, 3, 1, 1, 3, 2 };
+        static constexpr std::uint32_t vertex_count = 4;
+        static constexpr std::uint32_t index_count = 6;
+        static constexpr std::uint32_t indices[index_count] = { 0, 3, 1, 1, 3, 2 };
 
         vertex_type vertices[vertex_count];
 
@@ -31,9 +33,9 @@ namespace cppengine {
         vertex_type& operator[](int i);
         vertex_type const& operator[](int i) const;
 
-        std::size_t constexpr inline getVertexCount() const noexcept { return vertex_count; }
-        constexpr inline std::size_t const * getIndices() const noexcept { return indices; }
-        constexpr inline std::size_t getIndexCount() const noexcept { return index_count; }
+        std::uint32_t constexpr inline getVertexCount() const noexcept { return vertex_count; }
+        constexpr inline std::uint32_t const * getIndices() const noexcept { return indices; }
+        constexpr inline std::uint32_t getIndexCount() const noexcept { return index_count; }
 
     };
