+++ glfw_glad_integration/include/glad_model_handle.hpp
@@ -1,5 +1,6 @@
 #ifndef GLAD_MODEL_HANDLE_HPP
 #define GLAD_MODEL_HANDLE_HPP
+#include <cstdint>
 #include <string>
 
 #include "graphics/model_handle.hpp"
@@ -11,22 +12,22 @@ namespace cppengine {
     private:
         std::string name;
         GLuint vao, vbo, ebo;
-        std::size_t elementCount;
+        std::uint32_t elementCount;
 
     public:
         GladModelHandle(
             std::string const &name,
             Vector3 const *vertices,
-            std::size_t vertexCount,
-            std::size_t const *indices,
-            std::size_t indexCount);
+            std::uint32_t vertexCount,
+            std::uint32_t const *indices,
+            std::uint32_t indexCount);
 
         GladModelHandle(
             std::string &&name,
             Vector3 const *vertices,
-            std::size_t vertexCount,
-            std::size_t const *indices,
-            std::size_t indexCount);
+            std::uint32_t vertexCount,
+            std::uint32_t const *indices,
+            std::uint32_t indexCount);
 
         GladModelHandle(GladModelHandle const &) = delete;
         GladModelHandle &operator=(GladModelHandle const &) = delete;
@@ -35,7 +36,7 @@ namespace cppengine {
         void unbindModel() override;
 
         GLuint getId() const;
-        std::size_t getElementCount() const;
+        std::uint32_t getElementCount() const;
         std::string const &getName() const override;
 
         ~GladModelHandle() override;
